thread: Fix the barrier in threads
Currently, the barrier in the thread feature has an unmatched count
with the number of the thread group. I would expect that the barrier
should provide cluster-level synchronization with the cluster object.
So, This patch fixes the above problems and adapts the barrier
mechanism in the thread feature to allow cluster-level
synchronization.